backend/requirements.txt
Flask
youtube-transcript-api
transformers
flask-cors
tensorflow
flask_wtf
bert-extractive-summarizer